Improve handling of invalid regular expressions
If the user types a regular expression that is not understood by
Emacs, that doesn't mean that it's not going to be understood by "git
grep".  In any case, we don't want to propagate the error up to Helm /
Emacs just because we couldn't highlight some part of the string.